Fix multiple hits not showing all the info
Summary:
We noticed that not alertable hits were losing info because
their triggered hit time was not being recorded so later info
that followed wasnt being recorded.

Test Plan:
Run local console. Trigger a hit. Check info. Trigger same token
again quickly afterwards. Check that you got a mail for the first hit
and not the second one, but the second hit is recorded with all the
info

Reviewers: #devs, marco

Reviewed By: #devs, marco

Subscribers: marco

Differential Revision: http://phabricator.thinkst.com:9000/D955