Fail gracefully on browsers with no WebRTC support
Notably IE and Edge don't currently support WebRTC. The previous code
fails with the pc object not having createDataChannel defined. This would
not be a problem, but for it stopping subsequent JS code running. And
in the  case of the HTTP Slow Redirect token, the redirect in JS never runs.