Add support for a Generic DNS encoding
The DNS token can be used to transmit limited but arbitrary data, which
will be included in the incident report.

Any DNS token matching this format will be decoded as generic data:

 <base32 string with 63-byte labels>\.G[0-9]{2}\.<DNS token>

The bytes preceding '\.G[0-9]{2}\.' are stripped of their periods, then
a Base32 decode is attempted. If usccessful, the incident has a new
field added to the 'src_data' dictionary, named 'generic_data'.

This will appear in notification emails under Additional Details as
Generic Data.

A few notes to keep in mind:
 - To use this, you'll need to encode the data yourself into a DNS token
   you've previously generated.
 - When encoding, the normal DNS hostname format rules apply. Total
   length of the hostname (i.e. encoded data, plus the DNS token suffix)
   can't exceed 253 bytes, and no single label can be longer than 63
   bytes.
 - Case is often not preserved through intermediate DNS servers, so
   don't worry about getting it right. The server upper cases everything
   first.