Filter out spam of log errors coming through
After running canarytokens.org with the new ability to post error level
log lines to a webhook, we noticed that there were a couple of message
types that were coming through as spam. These messages were not helpful
at all to receive. The fixes range from simply changing the log level to
`warn` for the bitcoin and imgur spam. To fixing a bug in the DNS
channel code that would lead to a:
`exceptions.TypeError: 'NoneType' object has no attribute '__getitem__'`
traceback whenever the src_ip was not correctly handled.
There are also 2x new direct string matching filters in the loghandlers
code that takes away two separate spam cases that we do not believe
needs to be sent to the webhook as they are not actionable.