Added support for MS Excel
Embed a canarytoken image into an Excel workbook.

The image is loaded from a URL that is fetched on open. This is almost
identical to the Word functionality, just some changes to work with an excel
template file.

I used aspects of the word template metadata for the excel template.

One note, the image is in a random "far" cell (col:61 row:397), but a large
copy paste could overwrite it. It could be moved to a "further" cell.

Finally, this relies on the image being a 1x1 transparent as it will render any
actual picture content. Additional mandatory minimization could be achieved in
drawing1.xml but I haven't looked into those yet.

I only tested this on a single instance of Excel for Mac.