Add MySQL dump token
This diff adds Jacob's MySQL token to Canarytokens.org.

Notable things for this token is that it introduces a new
listen channel on which we alert on.

A Twisted service listens on port 3306 and responds with MySQL
server info. This is needed in order for the Replica flow to trigger
on the MySQL client.

This token currently grabs the attacker's IP, hostname and MySQL locale.