Add a Network Folder Canarytoken
This new Canarytoken type gives users the ability to mount a WebDAV
endpoint, with unique credentials. If the WebDAV mount is browsed and
the file are access, we send a notification.

It relies on a WebDAV server written separately that is hosted as a
Cloudflare Worker. This feature will not run on self-hosted Canarytokens
instances.

Actual author: Jacob Torrey