Fix alert for Web BUG POST requests
The POST request hanling for the following token types:
  - WEB
  - WEB_IMAGE
  - SLOW_REDIRECT

were very particular about the arguments which are alowed in the 'post'
request. The motivation for this is not entirely clear, but it seems
like it may want to strip out certain fields if key with the name 'key'
is present in the request arguments. This is fine, but if this is not
met then the POST request is dropped. Instead of dropping it, just let
'request_GET' handle it.