tests: unit: Add a unit test for POST requests
This is based on the test_channel_http_GET_and_POST_back test, though
it only does a POST request, and checks that the token still responds.

It asserts that the 'request_args' fields in the responds matches that
of the arguments set during the request.

Also - sneak in a small change to add 'WEB_IMAGE' to the normal
'test_channel_http_GET' test.