workflows: cypress: Change 'push' trigger to 'workflow_call' (#793)
The cypress tests do not build or deploy anything, they assume that the
version to be tested is already deployed. This means that they need to
depend on the deploy being finished, instead of running immediately on
'push'.

This adds two extra jobs in the build workflow. One as a gate which
sets a variable if 'any' of the deploy steps succeeded, and then one to
trigger the cypress tests if the test-gate job succeeded.

In the cypress workflow the 'on push' trigger is replaced with a
'workflow_dispatch' and a 'workflow_call' trigger. The 'workflow_call'
allows for it to be called with a 'uses:' step from build.yaml, and the
'workflow_dispatch' allows for a manual trigger from the github UI.