Respect `CANARY_MAX_HISTORY` and make the hit limit consistent (#792)
Canarytokens [historically had a setting](https://github.com/thinkst/canarytokens-docker/wiki/Setting-up-your-Canarytokens-DNS-correctly#:~:text=CANARY_MAX_HISTORY) for the maximum number of hits tracked for a given token. This PR restores that config option, `CANARY_MAX_HISTORY`, and gives it a default value of 50.

Additionally, it resolves a bug whereby a different number of hits were kept depending on the HTTP method most recently used to trigger the token, and adds a unit test to ensure this no longer happens.